Moved SumWhenTemplateSpec and subclasses into its own file
